requirements.txt
torch>=2.0.0
transformers>=4.30.0
einops>=0.6.0
numpy>=1.24.0
Pillow>=9.0.0
pytorch-lightning>=2.0.0
torchvision>=0.15.0
albumentations>=1.3.0
wandb>=0.15.0
opencv-python>=4.8.0
tqdm>=4.65.0
